+++ src/IdentityManager.php
@@ -65,9 +65,15 @@ class IdentityManager implements IdentityManagerInterface
      */
     private function checkEmailOrCreateUser(string $email): void
     {
+        // remove never-logged-in users older than 15 minutes (could be triggered by cron instead)
+        $this->executeWriteQuery(
+            "DELETE u FROM `{$this->tablePrefix}users` AS u WHERE u.last_login IS NULL"
+            . " AND u.created < (CURRENT_TIMESTAMP - INTERVAL 15 MINUTE)"
+            . " AND NOT EXISTS (SELECT 1 FROM `{$this->tablePrefix}session_user` AS su"
+            . " WHERE su.user_id = u.id LIMIT 1);"
+        );
         // Validate existence of the user or create it
         // Select email From users and if nothing returned, then INSERT email INTO users
-        // TODO destroy never-logged-in users older than 15 minutes.
         // Validate email format. Throwing the generic InvalidArgumentException from Webmozart is acceptable
         // for now; tests can catch it specifically if desired.
         Assert::email($email);